Fix #10349: Move getter methods from topic_services to topic_fetchers (#11861)
