+++ core/domain/classroom_services_test.py
@@ -20,7 +20,7 @@ from __future__ import unicode_literals  # pylint: disable=import-only-modules
 from constants import constants
 from core.domain import classroom_services
 from core.domain import config_services
-from core.domain import topic_services
+from core.domain import topic_fetchers
 from core.tests import test_utils
 
 
@@ -34,7 +34,7 @@ class ClassroomServicesTests(test_utils.GenericTestBase):
         self.set_admins([self.ADMIN_USERNAME, self.ADMIN_USERNAME])
 
     def test_can_get_classroom_by_url_fragment(self):
-        topic_id = topic_services.get_new_topic_id()
+        topic_id = topic_fetchers.get_new_topic_id()
         config_services.set_property(
             self.user_id_admin, 'classroom_pages_data', [{
                 'name': 'math',
@@ -53,7 +53,7 @@ class ClassroomServicesTests(test_utils.GenericTestBase):
         self.assertIsNone(classroom)
 
     def test_get_classroom_url_fragment_for_topic_id(self):
-        topic_id = topic_services.get_new_topic_id()
+        topic_id = topic_fetchers.get_new_topic_id()
         config_services.set_property(
             self.user_id_admin, 'classroom_pages_data', [{
                 'name': 'math',
@@ -68,7 +68,7 @@ class ClassroomServicesTests(test_utils.GenericTestBase):
         self.assertEqual(classroom_url_fragment, 'math-one')
 
     def test_return_default_if_associated_classroom_is_not_found(self):
-        topic_id = topic_services.get_new_topic_id()
+        topic_id = topic_fetchers.get_new_topic_id()
         config_services.set_property(
             self.user_id_admin, 'classroom_pages_data', [{
                 'name': 'math',
