+++ core/domain/topic_fetchers_test.py
@@ -44,7 +44,7 @@ class TopicFetchersUnitTests(test_utils.GenericTestBase):
 
     def setUp(self):
         super(TopicFetchersUnitTests, self).setUp()
-        self.TOPIC_ID = topic_services.get_new_topic_id()
+        self.TOPIC_ID = topic_fetchers.get_new_topic_id()
         changelist = [topic_domain.TopicChange({
             'cmd': topic_domain.CMD_ADD_SUBTOPIC,
             'title': 'Title',
@@ -156,7 +156,7 @@ class TopicFetchersUnitTests(test_utils.GenericTestBase):
         self.assertEqual(topic.to_dict(), expected_topic)
 
     def test_get_topic_by_version(self):
-        topic_id = topic_services.get_new_topic_id()
+        topic_id = topic_fetchers.get_new_topic_id()
         self.save_new_topic(
             topic_id, self.user_id, name='topic name',
             abbreviated_name='topic-name', url_fragment='topic-name',
@@ -201,7 +201,7 @@ class TopicFetchersUnitTests(test_utils.GenericTestBase):
             self.user_admin, self.user_a,
             topic_domain.ROLE_MANAGER, self.TOPIC_ID)
 
-        topic_rights = topic_services.get_topic_rights_with_user(self.user_id_a)
+        topic_rights = topic_fetchers.get_topic_rights_with_user(self.user_id_a)
 
         self.assertEqual(len(topic_rights), 1)
         self.assertEqual(topic_rights[0].id, self.TOPIC_ID)
@@ -214,3 +214,94 @@ class TopicFetchersUnitTests(test_utils.GenericTestBase):
         self.assertEqual(topic_commit_log_entry.commit_type, 'create')
         self.assertEqual(topic_commit_log_entry.topic_id, self.TOPIC_ID)
         self.assertEqual(topic_commit_log_entry.user_id, self.user_id)
+
+    def test_get_all_summaries(self):
+        topic_summaries = topic_fetchers.get_all_topic_summaries()
+
+        self.assertEqual(len(topic_summaries), 1)
+        self.assertEqual(topic_summaries[0].name, 'Name')
+        self.assertEqual(topic_summaries[0].canonical_story_count, 0)
+        self.assertEqual(topic_summaries[0].additional_story_count, 0)
+        self.assertEqual(topic_summaries[0].total_skill_count, 2)
+        self.assertEqual(topic_summaries[0].uncategorized_skill_count, 2)
+        self.assertEqual(topic_summaries[0].subtopic_count, 1)
+
+    def test_get_multi_summaries(self):
+        topic_summaries = topic_fetchers.get_multi_topic_summaries([
+            self.TOPIC_ID, 'invalid_id'])
+
+        self.assertEqual(len(topic_summaries), 2)
+        self.assertEqual(topic_summaries[0].name, 'Name')
+        self.assertEqual(topic_summaries[0].description, 'Description')
+        self.assertEqual(topic_summaries[0].canonical_story_count, 0)
+        self.assertEqual(topic_summaries[0].additional_story_count, 0)
+        self.assertEqual(topic_summaries[0].total_skill_count, 2)
+        self.assertEqual(topic_summaries[0].uncategorized_skill_count, 2)
+        self.assertEqual(topic_summaries[0].subtopic_count, 1)
+        self.assertIsNone(topic_summaries[1])
+
+    def test_get_all_skill_ids_assigned_to_some_topic(self):
+        change_list = [topic_domain.TopicChange({
+            'cmd': topic_domain.CMD_MOVE_SKILL_ID_TO_SUBTOPIC,
+            'old_subtopic_id': None,
+            'new_subtopic_id': 1,
+            'skill_id': self.skill_id_1
+        })]
+        topic_services.update_topic_and_subtopic_pages(
+            self.user_id_admin, self.TOPIC_ID, change_list,
+            'Moved skill to subtopic.')
+        topic_id = topic_fetchers.get_new_topic_id()
+        self.save_new_topic(
+            topic_id, self.user_id, name='Name 2', description='Description',
+            abbreviated_name='random', url_fragment='name-three',
+            canonical_story_ids=[], additional_story_ids=[],
+            uncategorized_skill_ids=[self.skill_id_1, 'skill_3'],
+            subtopics=[], next_subtopic_id=1)
+        self.assertEqual(
+            topic_fetchers.get_all_skill_ids_assigned_to_some_topic(),
+            {self.skill_id_1, self.skill_id_2, 'skill_3'})
+
+    def test_get_topic_summary_from_model(self):
+        topic_summary_model = topic_models.TopicSummaryModel.get(self.TOPIC_ID)
+        topic_summary = topic_fetchers.get_topic_summary_from_model(
+            topic_summary_model)
+
+        self.assertEqual(topic_summary.id, self.TOPIC_ID)
+        self.assertEqual(topic_summary.name, 'Name')
+        self.assertEqual(topic_summary.description, 'Description')
+        self.assertEqual(topic_summary.canonical_story_count, 0)
+        self.assertEqual(topic_summary.additional_story_count, 0)
+        self.assertEqual(topic_summary.uncategorized_skill_count, 2)
+        self.assertEqual(topic_summary.total_skill_count, 2)
+        self.assertEqual(topic_summary.subtopic_count, 1)
+        self.assertEqual(topic_summary.thumbnail_filename, 'topic.svg')
+        self.assertEqual(topic_summary.thumbnail_bg_color, '#C6DCDA')
+
+    def test_get_topic_summary_by_id(self):
+        topic_summary = topic_fetchers.get_topic_summary_by_id(self.TOPIC_ID)
+
+        self.assertEqual(topic_summary.id, self.TOPIC_ID)
+        self.assertEqual(topic_summary.name, 'Name')
+        self.assertEqual(topic_summary.description, 'Description')
+        self.assertEqual(topic_summary.canonical_story_count, 0)
+        self.assertEqual(topic_summary.additional_story_count, 0)
+        self.assertEqual(topic_summary.uncategorized_skill_count, 2)
+        self.assertEqual(topic_summary.subtopic_count, 1)
+        self.assertEqual(topic_summary.thumbnail_filename, 'topic.svg')
+        self.assertEqual(topic_summary.thumbnail_bg_color, '#C6DCDA')
+
+    def test_get_new_topic_id(self):
+        new_topic_id = topic_fetchers.get_new_topic_id()
+
+        self.assertEqual(len(new_topic_id), 12)
+        self.assertEqual(topic_models.TopicModel.get_by_id(new_topic_id), None)
+
+    def test_get_multi_rights(self):
+        topic_rights = topic_fetchers.get_multi_topic_rights([
+            self.TOPIC_ID, 'invalid_id'])
+
+        self.assertEqual(len(topic_rights), 2)
+        self.assertEqual(topic_rights[0].id, self.TOPIC_ID)
+        self.assertEqual(topic_rights[0].manager_ids, [])
+        self.assertFalse(topic_rights[0].topic_is_published)
+        self.assertIsNone(topic_rights[1])
